requirements.txt
gTTS==2.2.2
discord.py[voice]==1.7.3
